requirements.txt
openai
google-generativeai
speechrecognition
pyttsx3
pyautogui
keyboard
selenium
requests